+++ ohsome_quality_api/api/api.py
@@ -3,7 +3,7 @@ import logging
 import os
 from typing import Any, Union
 
-from fastapi import FastAPI, HTTPException, Request, status
+from fastapi import Depends, FastAPI, HTTPException, Request, status
 from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
@@ -24,6 +24,7 @@ from ohsome_quality_api import (
     __version__,
     oqt,
 )
+from ohsome_quality_api.api.request_context import set_request_context
 from ohsome_quality_api.api.request_models import (
     AttributeCompletenessFilterRequest,
     AttributeCompletenessKeyRequest,
@@ -70,15 +71,10 @@ from ohsome_quality_api.utils.exceptions import (
     OhsomeApiError,
     SizeRestrictionError,
     TopicDataSchemaError,
-    ValidationError,
 )
 from ohsome_quality_api.utils.helper import (
     get_class_from_key,
     json_serialize,
-)
-from ohsome_quality_api.utils.validators import (
-    validate_attribute_topic_combination,
-    validate_indicator_topic_combination,
 )
 
 MEDIA_TYPE_GEOJSON = "application/geo+json"
@@ -114,6 +110,7 @@ app = FastAPI(
     openapi_tags=TAGS_METADATA,
     docs_url=None,
     redoc_url=None,
+    dependencies=[Depends(set_request_context)],
 )
 
 app.add_middleware(
@@ -181,13 +178,8 @@ async def validation_exception_handler(
 @app.exception_handler(TopicDataSchemaError)
 @app.exception_handler(OhsomeApiError)
 @app.exception_handler(SizeRestrictionError)
-@app.exception_handler(ValidationError)
 async def custom_exception_handler(
-    _: Request,
-    exception: TopicDataSchemaError
-    | OhsomeApiError
-    | SizeRestrictionError
-    | ValidationError,
+    _: Request, exception: TopicDataSchemaError | OhsomeApiError | SizeRestrictionError
 ):
     """Exception handler for custom exceptions."""
     return JSONResponse(
@@ -272,13 +264,6 @@ async def post_attribute_completeness(
     parameters: AttributeCompletenessKeyRequest | AttributeCompletenessFilterRequest,
 ) -> Any:
     """Request the Attribute Completeness indicator for your area of interest."""
-    if isinstance(parameters, AttributeCompletenessKeyRequest):
-        for attribute in parameters.attribute_keys:
-            validate_attribute_topic_combination(
-                attribute,
-                parameters.topic,
-            )
-
     return await _post_indicator(request, "attribute-completeness", parameters)
 
 
@@ -300,9 +285,7 @@ async def post_attribute_completeness(
     },
 )
 async def post_indicator(
-    request: Request,
-    key: IndicatorEnumRequest,
-    parameters: IndicatorRequest,
+    request: Request, key: IndicatorEnumRequest, parameters: IndicatorRequest
 ) -> Any:
     """Request an indicator for your area of interest."""
     return await _post_indicator(request, key.value, parameters)
@@ -313,7 +296,6 @@ async def _post_indicator(
     key: str,
     parameters: IndicatorRequest,
 ) -> Any:
-    validate_indicator_topic_combination(key, parameters.topic)
     indicators = await oqt.create_indicator(key=key, **dict(parameters))
 
     if request.headers["accept"] == MEDIA_TYPE_JSON: